chore : move CheckBox file
